+++ packages/akashic-cli-serve/src/client/operator/DevtoolOperator.ts
@@ -61,4 +61,20 @@ export class DevtoolOperator {
 		// (console からグローバルに格納できる Chrome などでは不要なので暫定)
 		(window as any).__testbed.$0 = e;
 	}
+
+	volumeChangeTo = (vol: number): void => {
+		this.store.devtoolUiStore.volumeSeekTo(vol);
+		const atsumaruApi = (window as any).RPGAtsumaru;
+		if (atsumaruApi) {
+			atsumaruApi.volumeTrigger.fire(vol / 100);
+		}
+	}
+
+	volumeSeekTo = (vol: number): void => {
+		this.store.devtoolUiStore.endVolumeSeek(vol);
+		const atsumaruApi = (window as any).RPGAtsumaru;
+		if (atsumaruApi) {
+			atsumaruApi.volumeTrigger.fire(vol / 100);
+		}
+	}
 }